First Code Commit. Several functions for interfacing with spotify.
Allows for the user to authenticate if they have the environment
variables set appropriately. Authentication works. It then is capable of
displaying the playlists (public and private) of that user, and the
number of tracks. The user can select which playlist they want to
modify.

Current bugs result from trying to show the tracklist. Getting an HTTP
error, not sure what the problem is.

Added .gitignore for a cache file that shows up for the spotify user.